Revert "Feature/docs"
