--- apps/app/src/components/index.ts
@@ -1,1 +1,0 @@
-export {}
