--- apps/app/src/components/sidebar-menu/index.ts
@@ -1,3 +1,0 @@
-import SidebarMenu from "./SidebarMenu";
-
-export default SidebarMenu
