+++ apps/app/src/index.css
@@ -1,10 +1,7 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
 :root {
-  font-family: 'Roboto', sans-serif;
+  font-family: Inter, Avenir, Helvetica, Arial, sans-serif;
   font-size: 16px;
+  line-height: 24px;
   font-weight: 400;
 
   font-synthesis: none;
