--- apps/app/src/locale/index.ts
@@ -1,1 +1,0 @@
-export {}
