--- apps/app/src/pages/error-page/index.ts
@@ -1,3 +1,0 @@
-import ErrorPage from "./ErrorPage";
-
-export default ErrorPage
