--- apps/app/src/pages/overview/index.ts
@@ -1,3 +1,0 @@
-import Overview from "./Overview";
-
-export default Overview
