--- apps/app/src/styles/_mixin.scss
@@ -1,66 +1,0 @@
-@mixin materialStyle($inputHeight) {
-  $labelHeight: 22px;
-
-  .ant-form-item {
-    margin-bottom: 0;
-
-    .ant-form-item-label {
-      z-index: 1;
-      height: 10px;
-      overflow: visible;
-
-      label {
-        top: calc(10px + ($inputHeight - $labelHeight) / 2);
-        left: 6px;
-        transform-origin: top left;
-        transition: all 0.2s;
-        padding: 0 0.4rem;
-        background-color: white;
-        color: gray;
-        height: $labelHeight;
-
-        &.ant-form-item-required::before {
-          display: none;
-        }
-
-        &.ant-form-item-required::after {
-          display: inline-block;
-          margin: 0;
-          margin-inline-start: 3px;
-          color: inherit;
-          content: "*";
-        }
-      }
-    }
-
-    &:has(input:focus) .ant-form-item-label label,
-    &:has(input[value]:not([value=''])) .ant-form-item-label label,
-    &:has(textarea:focus) .ant-form-item-label label,
-    &:has(textarea:not(:empty)) .ant-form-item-label label {
-      transform: translateY(- calc($inputHeight / 2)) scale(0.75);
-    }
-
-    &:has(input:hover) .ant-form-item-label label,
-    &:has(label:hover) .ant-form-item-label label,
-    &:has(.ant-input-number-handler-wrap:hover) .ant-form-item-label label,
-    &:has(.ant-picker:hover) .ant-form-item-label label,
-    &:has(textarea:hover) .ant-form-item-label label,
-    &:has(input:focus) .ant-form-item-label label,
-    &:has(textarea:focus) .ant-form-item-label label,
-    &:has(.ant-picker-focused) .ant-form-item-label label {
-      color: $primary-color;
-    }
-
-    &:has(label:hover) .ant-form-item-control .ant-select .ant-select-selector,
-    &:has(label:hover) .ant-form-item-control .ant-input-number,
-    &:has(label:hover) .ant-form-item-control .ant-picker {
-      border-color: #ffbd52;
-    }
-
-    @media (max-width: 575px) {
-      .ant-form-item-label {
-        flex: none;
-      }
-    }
-  }
-}
