--- apps/app/src/styles/_variable.scss
@@ -1,1 +1,0 @@
-$primary-color: #FEA628;
